Move resource out to its own file
